Move gNMI-1.2 test out of an ate_tests directly
Move the full config replace benchmarking tests out of an ATE API
specific directory. The test does rely on the internal/setup/setup.go
test that contains ATE configuration code, but no ATE configuration
is actually done by this test.
