+++ feature/experimental/system/gnmi/benchmarking/tests/full_configuration_replace_test/metadata.textproto
